api/requirements.txt
# Vercel Serverless Requirements (minimal)
google-generativeai==0.3.2
Pillow==10.1.0
